+++ paleomix/tools/bam_pipeline/config.py
@@ -88,8 +88,6 @@ def parse_config(argv, pipeline_variant):
                      help = "May be specified one or more times with options to be passed "
                             "tot the JRE (Jave Runtime Environment); e.g. to change the "
                             "maximum amount of memory (default is -Xmx4g)")
-    group.add_option("--to-dot-file", dest="dot_file",
-                     help="Write dependency tree to the specified dot-file.")
     parser.add_option_group(group)
 
     return per_host_cfg.parse_args(parser, argv)