+++ paleomix/tools/phylo_pipeline/pipeline.py
@@ -125,11 +125,6 @@ def main(argv):
         logger.info("Printing required executables ...")
         pipeline.print_required_executables()
         return 0
-    elif config.dot_file:
-        logger.info("Writing dependency graph to %r ...", config.dot_file)
-        if not pipeline.to_dot(config.dot_file):
-            return 1
-        return 0
 
     if not pipeline.run(max_threads=config.max_threads,
                         dry_run=config.dry_run):