+++ paleomix/tools/zonkey/pipeline.py
@@ -77,9 +77,6 @@ def run_pipeline(config, nodes, msg):
     elif config.list_input_files:
         pipeline.print_input_files()
         return True
-    elif config.dot_file:
-        logger.info("Writing dependency graph to %r ...", config.dot_file)
-        return pipeline.to_dot(config.dot_file)
 
     return pipeline.run(max_threads=config.max_threads,
                         dry_run=config.dry_run)